with-django/requirements.txt
django==5.1.4
python-dotenv==1.0.1
polar-sdk==0.13.3
svix==1.44.0
